Emulate SkImageFilters::Xfermode with SkImageFilters::Blend
The release notes (below) seems to be mis-information: while the two take identical argument types,
skia.BlendMode.kSrc was valid input for SkImageFilters::Xfermode but does not result in valid
filter for SkImageFilters::Blend. The changed test value to kSrcIn (in another commit) is from
upstream's diff. Other might be acceptable, but kSrc is not. More investigation might be needed for differences
in acceptable inputs.

- Milestone 88
  ------------
    * Add new SkImageFilters::Blend factory function, in place of the now deprecated
    SkImageFilters::Xfermode factory function. Behavior is identical, but name better matches
    conventions in SkShader and SkColorFilter.
    https://review.skia.org/324623